requirements.txt
# Core dependencies
torch>=1.8.0
transformers>=4.10.0
scikit-learn>=0.24.0
pandas>=1.2.0
numpy>=1.19.0
matplotlib>=3.3.0
seaborn>=0.11.0

# SOTA model dependencies
python-dotenv>=0.19.0
requests>=2.25.0
tqdm>=4.62.0
librosa>=0.9.0
soundfile>=0.10.0
pydub>=0.25.0

# For audio processing
ffmpeg-python>=0.2.0

# For OpenRouter integration
openai>=1.0.0